requirements.txt
streamlit
openai
requests
pyyaml
python-dotenv
